protect createReviw, user update info, location creation with auth
