+++ application/src/app/api/users/route.ts
@@ -2,6 +2,8 @@ import { NextRequest, NextResponse } from "next/server";
 import { Role } from "@prisma/client";
 import bcrypt from "bcryptjs";
 import prisma from "../../../../prisma/prisma";
+import { getServerSession } from "next-auth";
+import { authOptions } from "../auth/[...nextauth]/authOptions";
 
 export const dynamic = 'force-dynamic'
 
@@ -38,7 +40,6 @@ export async function GET() {
   }
 }
 
-
 /**
  * @Endpoint - POST /api/users
  * @param - See Interface UserData
@@ -49,11 +50,10 @@ export async function POST(req: NextRequest) {
   try {
     // Get all data from the request body using the format we defined in the interface
     const body: UserData = await req.json();
-    
+
     // role is not required. If not provided, default to 'USER'
     const requiredFields: (keyof UserData)[] = [
       'username',
-      'email',
       'password'
     ];
 
@@ -104,6 +104,93 @@ export async function POST(req: NextRequest) {
     );
   } catch (error: any) {
     console.log(`[ERROR]: Error in POST of api/users/route.ts: ${error}`);
+    return NextResponse.json({ error: "Internal Server Error." }, { status: 500 });
+  }
+}
+
+/**
+ * @Auth - Required
+ * @Endpoint - PATCH /api/users/
+ * @description - Edit a user's data (email, password, settings)
+ * @returns - the user that was edited.
+ */
+export async function PATCH(req: NextRequest) {
+  console.log("PATCH request received");
+  const session = await getServerSession(authOptions);
+
+  // check if user is logged in
+  if (!session) {
+    return NextResponse.json({ error: "Unauthorized." }, { status: 401 });
+  }
+
+  try {
+
+    const { email, password, settings } = await req.json();
+    const userId = session.user.id;
+
+    if (!email && !password && !settings) {
+      return NextResponse.json({ error: "No email or password or settings provided." }, { status: 400 });
+    }
+
+    // check if user exists
+    const user = await prisma.user.findFirst({
+      where: {
+        id: userId
+      }
+    });
+
+    if (!user) {
+      return NextResponse.json({ error: "User does not exist." }, { status: 400 });
+    }
+
+    const updateData: any = {};
+
+    // if they provided a request to change emails
+    if (email) {
+      // check if email is already taken
+      const existingEmail = await prisma.user.findFirst({
+        where: { email },
+      });
+
+      if (existingEmail && existingEmail.id !== userId) {
+        return NextResponse.json({ error: "Email already taken by another user." }, { status: 400 });
+      }
+
+      // otherwise we can update the email
+      updateData.email = email;
+    }
+
+    // if they provided a request to change passwords
+    if (password) {
+      // hash the password before storing it in the database
+      const salt = bcrypt.genSaltSync(10);
+      const hashedPassword = bcrypt.hashSync(password, salt);
+      updateData.password = hashedPassword;
+    }
+
+    let updatedSettings = {};
+
+    // if updating their notifications or darkmode settings
+    if (settings) {
+      updatedSettings = await prisma.userSettings.update({
+        where: {
+          id: userId
+        },
+        data: settings
+      });
+    }
+
+    // finally update the user's email and password
+    const updatedUser = await prisma.user.update({
+      where: {
+        id: userId
+      },
+      data: updateData
+    });
+
+    return NextResponse.json(updatedUser);
+  } catch (error: any) {
+    console.log(`[ERROR]: Error in PATCH of api/users/route.ts: ${error}`);
     return NextResponse.json({ error: "Internal Server Error." }, { status: 500 });
   }
 }
