+++ application/src/app/prototype/README.md
@@ -2,4 +2,4 @@
 
 This folder serves as the backend team's folder for testing the connection between backend and frontend.
 
-Access the auth prototype test by changing the appending to the url /prototype/home or /prototype/auth/login
+Access the auth prototype test by appending to the url /prototype/home or /prototype/auth/login
